refactor: guiに関するコードと, parameterに関するコードを分離
