+++ src/canvases/modules/utils.ts
@@ -6,44 +6,6 @@ export type Point = {
   y: number
 }
 
-export type ParamNum = {
-  val: number,
-  min: number,
-  max: number,
-  isInt: boolean,
-}
-export const isParamNum = (arg: any): arg is ParamNum => {
-  return (typeof arg.val === 'number')
-    && (typeof arg.min === 'number') && (typeof arg.max === 'number')
-    && (typeof arg.isInt === 'boolean');
-}
-
-export type Color = {
-  r: number,
-  g: number,
-  b: number
-}
-export const isColor = (arg: any): arg is Color => {
-  return (typeof arg.r === 'number') && (typeof arg.g === 'number')
-    && (typeof arg.b === 'number');
-}
-export const codeToColor = (code: string): Color => {
-  code = code.replace('#', '');
-  code = code.toUpperCase();
-  console.log(code);
-  const c: Color = {
-    r: parseInt(code[0]+code[1], 16),
-    g: parseInt(code[2]+code[3], 16),
-    b: parseInt(code[4]+code[5], 16)
-  }
-  console.log(c);
-  return c;
-}
-
-export type ParamGUIs = {
-  [key: string]: p5.Element;
-}
-
 export const mod = (i: number, j: number): number => {
   return (i % j) < 0 ? (i % j) + 0 + (j < 0 ? -j : j) : (i % j + 0);
 }
@@ -67,36 +29,4 @@ export const nGon = (
   }
   p5.endShape(p5.CLOSE);
   p5.pop();
-}
-
-export const buildGUIs = (p5: p5, params: Object): ParamGUIs => {
-  let paramGUIs: ParamGUIs = {};
-  Object.entries(params).forEach(([key, value], i) => {
-    if (isParamNum(value)) {
-      let slider = p5.createSlider(
-        value.min, value.max, value.val,
-        value.isInt ? 1 : 0
-      );
-      slider.position(10, 30*i+10);
-      paramGUIs[key] = slider;
-    }else if (isColor(value)) {
-      let colorPicker = p5.createColorPicker(
-        p5.color(value.r, value.g, value.b)
-      );
-      colorPicker.position(10, 30*i+10);
-      paramGUIs[key] = colorPicker;
-    }
-  });
-  return paramGUIs;
-}
-
-export const updateGUIs = (params: Object, paramGUIs: ParamGUIs) => {
-  Object.entries(params).forEach(([key, value]) => {
-    if (isParamNum(value)) {
-      params[key].val = paramGUIs[key].value();
-    }else if (isColor(value)) {
-      let code: string = paramGUIs[key].value().toString();
-      params[key] = codeToColor(code);
-    }
-  });
 }
